Fixed #683 Fixed #684 Fixed #685
The flow backup feature saves flows with only the version name so only
one flow with the same version number is saved. 

Do a live backup of the database using H2's backup feature when the
backup process runs. 

On shutdown of Metl shutdown the H2 database by doing a "SHUTDOWN
COMPACT".