Fixed #654 & #655 AutoMap functionality update
Added ability to auto-map only filtered items on display as well as changed the fuzzy logic to not map every 'similar' name but will only auto-map the first similar attribute it finds and not all.  For example, if 'Name' is on source and 'Name1', 'Name2' and 'Name3' are on target only 'Name' will be auto-mapped to 'Name1' and not to all 3.